final done

+++ IPD_Income.cs
@@ -14,7 +14,7 @@ namespace Clinic_Management_System
     {
         databaseclass dbClass = new databaseclass();
         DataSet ds;
-        int totalProfit, totalCharge;
+        decimal totalProfit, totalCharge;
         public IPD_Income()
         {
             InitializeComponent();
@@ -36,8 +36,8 @@ namespace Clinic_Management_System
                 DataTable prescriptionTable = ds1.Tables[0];
                 DataTable combinedTable = new DataTable();
                 combinedTable.Columns.Add("profit_date", typeof(string));
-                combinedTable.Columns.Add("amount", typeof(int));
-                combinedTable.Columns.Add("total_pay", typeof(int));
+                combinedTable.Columns.Add("amount", typeof(decimal));
+                combinedTable.Columns.Add("total_pay", typeof(decimal));
 
                 int rowCount = Math.Max(profitTable.Rows.Count, prescriptionTable.Rows.Count);
                 for (int i = 0; i < rowCount; i++)
@@ -90,7 +90,7 @@ namespace Clinic_Management_System
             if (profitSet != null && profitSet.Tables.Count > 0)
             {
                 var totalProfitStr = profitSet.Tables[0].Rows[0]["total_profit"].ToString();
-                if (int.TryParse(totalProfitStr, out totalProfit))
+                if (decimal.TryParse(totalProfitStr, out totalProfit))
                 {
                     lblProfit.Text = "Today's Total Profit: " + totalProfit.ToString("C");
                 }
@@ -105,7 +105,7 @@ namespace Clinic_Management_System
             if (chargeSet != null && chargeSet.Tables.Count > 0)
             {
                 var totalChargeStr = chargeSet.Tables[0].Rows[0]["charge"].ToString();
-                if (int.TryParse(totalChargeStr, out totalCharge))
+                if (decimal.TryParse(totalChargeStr, out totalCharge))
                 {
                     lblCharge.Text = "Today's Total Charge: " + totalCharge.ToString("C");
                 }
@@ -136,7 +136,7 @@ namespace Clinic_Management_System
             if (totalProfitDS != null && totalProfitDS.Tables.Count > 0)
             {
                 var totalProfitStr = totalProfitDS.Tables[0].Rows[0]["total_profit"].ToString();
-                if (int.TryParse(totalProfitStr, out totalProfit))
+                if (decimal.TryParse(totalProfitStr, out totalProfit))
                 {
                     lblProfit.Text = "This Month's Total Profit: " + totalProfit.ToString("C");
                 }
@@ -153,7 +153,7 @@ namespace Clinic_Management_System
             if (chargeSet != null && chargeSet.Tables.Count > 0)
             {
                 var totalChargeStr = chargeSet.Tables[0].Rows[0]["charge"].ToString();
-                if (int.TryParse(totalChargeStr, out totalCharge))
+                if (decimal.TryParse(totalChargeStr, out totalCharge))
                 {
                     lblCharge.Text = "This Month's Total Charge: " + totalCharge.ToString("C");
                 }
@@ -181,7 +181,7 @@ namespace Clinic_Management_System
             if (totalProfitDS != null && totalProfitDS.Tables.Count > 0)
             {
                 var totalProfitStr = totalProfitDS.Tables[0].Rows[0]["total_profit"].ToString();
-                if (int.TryParse(totalProfitStr, out totalProfit))
+                if (decimal.TryParse(totalProfitStr, out totalProfit))
                 {
                     lblProfit.Text = "This Year's Total Profit: " + totalProfit.ToString("C");
                 }
@@ -197,7 +197,7 @@ namespace Clinic_Management_System
             if (chargeSet != null && chargeSet.Tables.Count > 0)
             {
                 var totalChargeStr = chargeSet.Tables[0].Rows[0]["charge"].ToString();
-                if (int.TryParse(totalChargeStr, out totalCharge))
+                if (decimal.TryParse(totalChargeStr, out totalCharge))
                 {
                     lblCharge.Text = "This Year's Total Charge: " + totalCharge.ToString("C");
                 }
